requirements.txt
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
gcloud==0.18.3
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
idna==3.6
importlib_metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
jwcrypto==1.5.6
MarkupSafe==2.1.5
oauth2client==4.1.3
oauthlib==3.2.2
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pyparsing==3.1.2
Pyrebase4==4.7.1
python-jwt==4.1.0
requests==2.29.0
requests-oauthlib==1.4.0
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.18
Werkzeug==3.0.1
zipp==3.18.1
